Remove legacy clients and update docs
