+++ output_kubernetes/rate_10.txt
