+++ output_kubernetes/rate_100.txt
