+++ source_code/config.json
@@ -1,8 +1,5 @@
 {
   "server": {
     "consoleLogging": false
-  },
-  "client": {
-    "consoleLogging": false
   }
 }
