+++ source_code/package.json
@@ -4,8 +4,7 @@
   "description": "Node.js server and client for autoscaling assignment",
   "main": "server.js",
   "scripts": {
-    "start": "node server.js",
-    "client": "node client.js"
+    "start": "node server.js"
   },
   "dependencies": {
     "axios": "^1.6.8",
